Added modularization and made correct imports
