+++ modules/__init__.py
@@ -1,0 +1,2 @@
+from .selenium_scrapping import selenium_simulation
+from .reviews_preprocessing import opinion_rating_compositor
